pyproject.toml
[project]
name = "ai-multiagent-restaurant-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.1",
    "langchain[anthropic]>=0.3.27",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
]
